Move DataSet to another folder
