+++ src/main/java/ru/naumen/perfhouse/influx/DBMock.java
@@ -1,6 +1,6 @@
 package ru.naumen.perfhouse.influx;
 
-import ru.naumen.sd40.log.parser.parsers.data.DataSet;
+import ru.naumen.sd40.log.parser.storages.DataSet;
 
 import java.util.HashMap;
