+++ src/main/java/ru/naumen/perfhouse/influx/IDataBase.java
@@ -1,6 +1,6 @@
 package ru.naumen.perfhouse.influx;
 
-import ru.naumen.sd40.log.parser.parsers.data.DataSet;
+import ru.naumen.sd40.log.parser.storages.DataSet;
 
 public interface IDataBase {
     void storeData(String dbName, long date, DataSet dataSet);
