+++ src/main/java/ru/naumen/sd40/log/parser/LogParser.java
@@ -8,6 +8,7 @@ import ru.naumen.sd40.log.parser.parsers.time.GCTimeParser;
 import ru.naumen.sd40.log.parser.parsers.time.ITimeParser;
 import ru.naumen.sd40.log.parser.parsers.time.SdngTimeParser;
 import ru.naumen.sd40.log.parser.parsers.time.TopTimeParser;
+import ru.naumen.sd40.log.parser.storages.DataSet;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
