+++ src/main/java/ru/naumen/sd40/log/parser/parsers/data/IDataParser.java
@@ -1,5 +1,7 @@
 package ru.naumen.sd40.log.parser.parsers.data;
 
+import ru.naumen.sd40.log.parser.storages.DataSet;
+
 public interface IDataParser {
     void parseLine(String line, DataSet dataSet);
 }
